+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * 1.11.7 - 2014-09-09:
     * Fix: Select2 errors on product, order and coupon pages.
     * Fix: Notice about `WYSIJA` constant.
+    * Improved: Add "Allow URL fopen" to System Info page.
+    * Improved: Handling of Jigoshop Settings scripts.
 * 1.11.6 - 2014-09-05:
     * Fix: PayPal invalid amounts.
     * Fix: JWOS with WordPress 4.0 compatibility.